Backend/requirements.txt
anyio==3.7.1
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.1.2
Automat==0.8.0
backports.zoneinfo==0.2.1
blinker==1.4
certifi==2019.11.28
cffi==1.15.1
channels==4.0.0
channels-auth-token-middlewares==1.1.0
channels-redis==4.1.0
chardet==3.0.4
click==8.1.3
cloud-init==22.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==40.0.1
cssselect==1.2.0
daphne==4.0.0
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==2.1.0
Django==4.2
django-cors-headers==4.2.0
django-debug-toolbar==4.0.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.93.4
drf-yasg==1.21.7
entrypoints==0.3
exceptiongroup==1.1.3
filelock==3.11.0
Flask==2.2.3
httplib2==0.14.0
hyperlink==21.0.0
idna==2.8
importlib-metadata==6.6.0
incremental==22.10.0
inflection==0.5.1
itemadapter==0.8.0
itemloaders==1.0.6
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
more-itertools==4.2.0
msgpack==1.0.5
netifaces==0.10.4
numpy==1.24.2
oauthlib==3.1.0
packaging==23.1
pandas==2.0.0
parsel==1.7.0
pexpect==4.6.0
Pillow==9.5.0
Protego==0.2.1
psycopg2==2.9.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
PyDispatcher==2.0.7
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==23.1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-decouple==3.8
pytz==2023.3
PyYAML==5.3.1
queuelib==1.6.2
redis==5.0.0
requests==2.22.0
requests-file==1.5.1
requests-unixsocket==0.2.0
Scrapy==2.8.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
snap==0.5
sniffio==1.3.0
sos==4.3
sqlparse==0.4.4
ssh-import-id==5.10
swagger-spec-validator==3.0.3
systemd-python==234
tldextract==3.4.0
twisted==23.8.0
txaio==23.1.1
typing-extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==27.10
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
w3lib==2.1.1
wadllib==1.3.3
watchgod==0.8.2
Werkzeug==2.2.3
zipp==1.0.0
zope.interface==6.0
